Groundwork for separate emulated/hooked controllers
