+++ HandOfLesserDriver/HandOfLesserDriver/src/controller_common.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <HandOfLesserCommon.h>
+#include "openvr_driver.h"
+
+namespace HOL::ControllerCommon
+{
+	extern vr::DriverPose_t generatePose(HOL::HandTransformPacket* packet, bool deviceConnected);
+
+
+
+}
