+++ HandOfLesserDriver/HandOfLesserDriver/src/device_provider.cpp
@@ -13,47 +13,10 @@ vr::EVRInitError MyDeviceProvider::Init(vr::IVRDriverContext* pDriverContext)
 	// OpenVR provides a macro to do this for us.
 	VR_INIT_SERVER_DRIVER_CONTEXT(pDriverContext);
 
-	// Let's add our controllers to the system.
-	// First, we need to actually instantiate our controller devices.
-	// We made the constructor take in a controller role, so let's pass their respective roles in.
-	my_left_controller_device_
-		= std::make_unique<ControllerDeviceDriver>(vr::TrackedControllerRole_LeftHand);
-	my_right_controller_device_
-		= std::make_unique<ControllerDeviceDriver>(vr::TrackedControllerRole_RightHand);
-
-	this->mTransport.init(9006); // Hardcoded for now, needs to be negotaited somehow
-	my_pose_update_thread_ = std::thread(&MyDeviceProvider::ReceiveDataThread, this);
-
-	// Now we need to tell vrserver about our controllers.
-	// The first argument is the serial number of the device, which must be unique across all
-	// devices. We get it from our driver settings when we instantiate, And can pass it out of the
-	// function with MyGetSerialNumber(). Let's add the left hand controller first (there isn't a
-	// specific order). make sure we actually managed to create the device. TrackedDeviceAdded
-	// returning true means we have had our device added to SteamVR.
-	if (!vr::VRServerDriverHost()->TrackedDeviceAdded(
-			my_left_controller_device_->MyGetSerialNumber().c_str(),
-			vr::TrackedDeviceClass_Controller,
-			my_left_controller_device_.get()
-		))
-	{
-		DriverLog("Failed to create left controller device!");
-		// We failed? Return early.
-		return vr::VRInitError_Driver_Unknown;
-	}
-
-	// Now, the right hand
-	// Make sure we actually managed to create the device.
-	// TrackedDeviceAdded returning true means we have had our device added to SteamVR.
-	if (!vr::VRServerDriverHost()->TrackedDeviceAdded(
-			my_right_controller_device_->MyGetSerialNumber().c_str(),
-			vr::TrackedDeviceClass_Controller,
-			my_right_controller_device_.get()
-		))
-	{
-		DriverLog("Failed to create right controller device!");
-		// We failed? Return early.
-		return vr::VRInitError_Driver_Unknown;
-	}
+	// Here we would add controllers, if we were defaulting to that
+	// For now we'll be defaulting to hooking existing controllers instead. 
+	// Lots of TODOs
+	this->mHandOfLesser.init();
 
 	return vr::VRInitError_None;
 }
@@ -84,106 +47,9 @@ bool MyDeviceProvider::ShouldBlockStandbyMode()
 void MyDeviceProvider::RunFrame()
 {
 
-	// call our devices to run a frame
-	if (my_left_controller_device_ != nullptr)
-	{
-		my_left_controller_device_->MyRunFrame();
-	}
 
-	if (my_right_controller_device_ != nullptr)
-	{
-		my_right_controller_device_->MyRunFrame();
-	}
-
-	// Now, process events that were submitted for this frame.
-	vr::VREvent_t vrevent{};
-	while (vr::VRServerDriverHost()->PollNextEvent(&vrevent, sizeof(vr::VREvent_t)))
-	{
-		if (my_left_controller_device_ != nullptr)
-		{
-			my_left_controller_device_->MyProcessEvent(vrevent);
-		}
-
-		if (my_right_controller_device_ != nullptr)
-		{
-			my_right_controller_device_->MyProcessEvent(vrevent);
-		}
-	}
 }
 
-void MyDeviceProvider::ReceiveDataThread()
-{
-	while (this->mActive)
-	{
-		HOL::NativePacket* rawPacket = this->mTransport.receive();
-		if (rawPacket == nullptr)
-		{
-			continue;
-		}
-		switch (rawPacket->packetType)
-		{
-		case HOL::NativePacketType::HandTransform: {
-			HOL::HandTransformPacket* packet = (HOL::HandTransformPacket*)rawPacket;
-
-			if (packet->valid)
-			{
-				ControllerDeviceDriver* controller;
-				if (packet->side == HOL::HandSide::LeftHand)
-				{
-					controller = this->my_left_controller_device_.get();
-				}
-				else
-				{
-					controller = this->my_right_controller_device_.get();
-				}
-
-				controller->UpdatePose(packet);
-				controller->SubmitPose();
-			}
-
-			break;
-		}
-
-		case HOL::NativePacketType::ControllerInput: {
-			HOL::ControllerInputPacket* packet = (HOL::ControllerInputPacket*)rawPacket;
-
-			if (packet->valid)
-			{
-				ControllerDeviceDriver* controller;
-				if (packet->side == HOL::HandSide::LeftHand)
-				{
-					controller = this->my_left_controller_device_.get();
-				}
-				else
-				{
-					controller = this->my_right_controller_device_.get();
-				}
-
-				controller->UpdateInput(packet);
-			}
-
-			break;
-		}
-
-		default: {
-			// Invalid packet type!
-		}
-		}
-	}
-
-	/*
-	while (is_active_)
-	{
-		// Inform the vrserver that our tracked device's pose has updated, giving it the pose
-	returned by our GetPose().
-		vr::VRServerDriverHost()->TrackedDevicePoseUpdated(my_controller_index_, GetPose(),
-	sizeof(vr::DriverPose_t));
-
-		// Update our pose every five milliseconds.
-		// In reality, you should update the pose whenever you have new data from your device.
-		std::this_thread::sleep_for(std::chrono::milliseconds(5));
-	}*/
-}
 
 //-----------------------------------------------------------------------------
 // Purpose: This function is called when the system enters a period of inactivity.
@@ -208,16 +74,5 @@ void MyDeviceProvider::LeaveStandby()
 //-----------------------------------------------------------------------------
 void MyDeviceProvider::Cleanup()
 {
-	// Let's join our pose thread that's running
-	// by first checking then setting is_active_ to false to break out
-	// of the while loop, if it's running, then call .join() on the thread
-	// if (is_active_.exchange(false))
-	this->mActive = false;
-	{
-		my_pose_update_thread_.join();
-	}
-
-	// Our controller devices will have already deactivated. Let's now destroy them.
-	my_left_controller_device_ = nullptr;
-	my_right_controller_device_ = nullptr;
+	this->mHandOfLesser.cleanup();
 }
